Move mobile-platform-capabilities to common
