+++ test/build-configurations.ts
@@ -15,7 +15,7 @@ import jsonSchemaConstantsLib = require("../lib/json-schema/json-schema-constant
 import stubs = require("./stubs");
 import yok = require("../lib/common/yok");
 import mobileHelperLib = require("../lib/common/mobile/mobile-helper");
-import mobilePlatformsCapabilitiesLib = require("../lib/mobile-platforms-capabilities");
+import mobilePlatformsCapabilitiesLib = require("../lib/common/appbuilder/mobile-platforms-capabilities");
 import devicePlatformsLib = require("../lib/common/mobile/device-platforms-constants");
 import hostInfoLib = require("../lib/common/host-info");
 import optionsLib = require("../lib/options");
